LayoutTests/imported/w3c/web-platform-tests/xhr/timeout-sync-expected.txt


PASS setting timeout attribute on sync request
PASS open() with async false when timeout is set
